if empty body, iOS crashed
When a get request with the content type of JSON has an empty body the application would crash. This should not happen, the developer it is forced to remove the content type which doesn't make sense. While on Android it would work fine. This at least brings consistency in that regard.